wip IgniteServerBase
