+++ modules/platforms/dotnet/Apache.Ignite.Tests/IgniteServerBase.cs
@@ -1,14 +1,79 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 namespace Apache.Ignite.Tests;
 
 using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Net;
 using System.Net.Sockets;
+using System.Threading;
+using System.Threading.Tasks;
 using Internal.Buffers;
 
 [SuppressMessage("Design", "CA1034:Nested types should not be visible", Justification = "Tests.")]
-public abstract class IgniteServerBase
+public abstract class IgniteServerBase : IDisposable
 {
+    private readonly Socket _listener;
+    private readonly CancellationTokenSource _cts = new();
+    private readonly object _disposeSyncRoot = new();
+    private volatile Socket? _handler;
+    private bool _disposed;
+
+    protected IgniteServerBase()
+    {
+        _listener = new Socket(IPAddress.Loopback.AddressFamily, SocketType.Stream, ProtocolType.Tcp);
+        _listener.NoDelay = true;
+
+        _listener.Bind(new IPEndPoint(IPAddress.Loopback, 0));
+        _listener.Listen(backlog: 1);
+        Task.Run(ListenLoop);
+    }
+
+    public void Dispose()
+    {
+        lock (_disposeSyncRoot)
+        {
+            if (_disposed)
+            {
+                return;
+            }
+
+            _cts.Cancel();
+            _handler?.Dispose();
+            _listener.Disconnect(false);
+            _listener.Dispose();
+            _cts.Dispose();
+
+            Dispose(true);
+            GC.SuppressFinalize(this);
+
+            _disposed = true;
+        }
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            // No-op.
+        }
+    }
+
     private static int ReceiveMessageSize(Socket handler)
     {
         using var buf = ReceiveBytes(handler, 4);
@@ -35,6 +100,40 @@ public abstract class IgniteServerBase
         return new PooledBuffer(buf, 0, size);
     }
 
+    private void ListenLoop()
+    {
+        while (!_cts.IsCancellationRequested)
+        {
+            try
+            {
+                ListenLoopInternal();
+            }
+            catch (Exception e)
+            {
+                if (e is SocketException or ConnectionLostException)
+                {
+                    continue;
+                }
+
+                Console.WriteLine("Error in FakeServer: " + e);
+            }
+        }
+    }
+
+    private void ListenLoopInternal()
+    {
+        while (!_cts.IsCancellationRequested)
+        {
+            using Socket handler = _listener.Accept();
+            _handler = handler;
+            handler.NoDelay = true;
+
+            // Read handshake.
+            using var magic = ReceiveBytes(handler, 4);
+            var msgSize = ReceiveMessageSize(handler);
+            using var handshake = ReceiveBytes(handler, msgSize);
+        }
+    }
 
     [SuppressMessage("Design", "CA1032:Implement standard exception constructors", Justification = "Tests.")]
     [SuppressMessage("Design", "CA1064:Exceptions should be public", Justification = "Tests.")]
